# OpenAPI/product-service.yml
openapi: 3.0.1
info:
  title: Product Service API
  version: "1.0"
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
  /change_batch_quantity:
    post:
      tags:
      - product-controller
      summary: Get allocations for order id
      operationId: ChangeBatchQuantity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeBatchQuantity"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/InvalidBatchReferenceException"
                - $ref: "#/components/schemas/InvalidSkuException"
  /allocate:
    post:
      tags:
      - product-controller
      summary: Allocate an order line to a batch
      operationId: allocate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Allocate"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Allocated"
                - $ref: "#/components/schemas/OutOfStockException"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidSkuException"
  /add_batch:
    post:
      tags:
      - product-controller
      summary: Add a batch to a sku
      operationId: addBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatch"
        required: true
      responses:
        "200":
          description: OK
  /allocations/{orderId}:
    get:
      tags:
      - product-controller
      summary: Get allocations for order id
      operationId: allocations
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    InvalidBatchReferenceException:
      type: object
      properties:
        cause:
          type: object
          properties:
            stackTrace:
              type: array
              items:
                type: object
                properties:
                  classLoaderName:
                    type: string
                  moduleName:
                    type: string
                  moduleVersion:
                    type: string
                  methodName:
                    type: string
                  fileName:
                    type: string
                  lineNumber:
                    type: integer
                    format: int32
                  className:
                    type: string
                  nativeMethod:
                    type: boolean
            message:
              type: string
            localizedMessage:
              type: string
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    InvalidSkuException:
      type: object
      properties:
        cause:
          type: object
          properties:
            stackTrace:
              type: array
              items:
                type: object
                properties:
                  classLoaderName:
                    type: string
                  moduleName:
                    type: string
                  moduleVersion:
                    type: string
                  methodName:
                    type: string
                  fileName:
                    type: string
                  lineNumber:
                    type: integer
                    format: int32
                  className:
                    type: string
                  nativeMethod:
                    type: boolean
            message:
              type: string
            localizedMessage:
              type: string
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ChangeBatchQuantity:
      type: object
      properties:
        batchRef:
          type: string
        qty:
          type: integer
          format: int32
    Allocated:
      type: object
      properties:
        orderId:
          type: string
        sku:
          type: string
        qty:
          type: integer
          format: int32
        batchRef:
          type: string
    OutOfStockException:
      type: object
      properties:
        cause:
          type: object
          properties:
            stackTrace:
              type: array
              items:
                type: object
                properties:
                  classLoaderName:
                    type: string
                  moduleName:
                    type: string
                  moduleVersion:
                    type: string
                  methodName:
                    type: string
                  fileName:
                    type: string
                  lineNumber:
                    type: integer
                    format: int32
                  className:
                    type: string
                  nativeMethod:
                    type: boolean
            message:
              type: string
            localizedMessage:
              type: string
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        sku:
          type: string
        message:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Allocate:
      type: object
      properties:
        orderId:
          type: string
        sku:
          type: string
        quantity:
          type: integer
          format: int32
    CreateBatch:
      type: object
      properties:
        reference:
          type: string
        sku:
          type: string
        qty:
          type: integer
          format: int32
        eta:
          type: string
          format: date
    Allocation:
      type: object
      properties:
        orderId:
          type: string
        sku:
          type: string
        batchRef:
          type: string